+++ .nginx/default.conf
@@ -1,0 +1,19 @@
+events {
+    worker_connections 1024;
+}
+
+http {
+    upstream fastapi_app {
+        server fastapi:8000;
+    }
+
+    server {
+        listen 80;
+        location / {
+            proxy_pass http://fastapi_app;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header Host $host;
+            proxy_redirect off;
+        }
+    }
+}
