+++ docker-compose.yml
@@ -5,11 +5,33 @@ services:
     image: postgres:latest
     container_name: database
     restart: always
-    ports:
-      - "5432:5432"
     environment:
       - POSTGRES_USER=${DB_USER}
       - POSTGRES_PASSWORD=${DB_PASSWORD}
       - POSTGRES_HOST=${DB_HOST}
       - POSTGRES_PORT=${DB_PORT}
       - POSTGRES_DB=${DB_NAME}
+    expose:
+      - 5432
+
+  fastapi:
+    build:
+      context: .
+      dockerfile: ./src/docker/Dockerfile.fastapi
+    container_name: fastapi
+    # env_file:
+    #   - .env
+    restart: always
+    expose:
+      - 8000
+
+  nginx:
+    image: nginx
+    container_name: nginx
+    restart: always
+    depends_on:
+      - fastapi
+    ports:
+      - "8080:8080"
+    volumes:
+      - .nginx/default.conf:/etc/nginx/nginx.conf:ro
