+++ src/app.py
@@ -2,7 +2,7 @@ from contextlib import asynccontextmanager
 from fastapi import FastAPI
 import aiohttp
 from .routers import hh_router
-from .database import create_tables, drop_tables
+from .config.database import create_tables, drop_tables
 
 
 aiohttp_clientsession: aiohttp.ClientSession = None
