+++ src/docker/Dockerfile.fastapi
@@ -1,0 +1,12 @@
+from python:3.12.4-slim
+
+LABEL name="fastapi"
+
+COPY ./src/req.txt /req.txt
+RUN python -m pip install -r ./req.txt
+
+WORKDIR /src
+COPY . .
+# COPY ./.env /.env
+CMD ["ls"]
+# CMD ["uvicorn", "src.app:app", "--host=0.0.0.0", "--port=8000"]
