+++ src/models.py
@@ -1,20 +1,7 @@
 from typing import Any
-from sqlalchemy.ext.asyncio import create_async_engine, async_sessionmaker
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 from sqlalchemy.dialects.postgresql import JSON
 
-engine = create_async_engine(
-    "postgresql+asyncpg://postgres:123@localhost:5432/"
-)
-new_session = async_sessionmaker(engine, expire_on_commit=False)
-
-async def create_tables():
-    async with engine.begin() as conn:
-       await conn.run_sync(Model.metadata.create_all)
-
-async def drop_tables():
-    async with engine.begin() as conn:
-        await conn.run_sync(Model.metadata.drop_all)
 
 class Model(DeclarativeBase):
     type_annotation_map = {
